requirements.txt
certifi
chardet
dnspython
idna
ipwhois
mysql-connector-python
protobuf
PySocks
requests
six
urllib3
